pyproject.toml
[project]
name = "bitoki"
version = "0.1.0"
description = "BTC/USD automated trading strategy with pattern recognition"
requires-python = ">=3.10"
dependencies = [
    "ccxt>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "ta>=0.11.0",
    "loguru>=0.7.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-login>=0.6.0",
    "flask-bcrypt>=1.0.0",
    "flask-mail>=0.9.0",
    "pyotp>=2.9.0",
    "qrcode>=7.4.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
